backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.10.6
pydantic-settings==2.1.0
python-dotenv==1.0.1
msal==1.26.0
requests==2.31.0
python-jose==3.3.0
python-docx==1.1.0
PyPDF2==3.0.1
